Adding RP page frontend
Once again, this was resurrected from the old project, but with
heavy modifications to conform to the new API, in addition to code
cleanup and such. It will eventually be a SPA which includes the
archive, stats, and export pages, but these are not yet
implemented. It does work as a crude solution for now, though.