Update API to be more RESTful
API now more closely matches the revised v1 API specification
from the documentation. It has been updated to be more RESTful.
It also now uses a single updateCounter rather than a msgCounter
and charaCounter when checking for updates. This counter is made
more explicitly available from the API calls. Finally, it is more
consistent in that it always uses "msgs" rather than a mix of
"messages" and "msgs"; "chara" is used instead of "character" too.

The client has been updated to work with these changes.