Add archive page to RP SPA. also bug fixes
got the #page-1 link to work correctely in the RP page SPA. It
now shows the first page of the archive, although page numbering
links are not yet present.

It also fixes a number of bugs in the API and in the frontend
related to switching between pages, fetching the correct
characters from the backend, displaying elements correctly on the
frontend, and other things.