Test suite now starts and stops the server.
We have a new suite of tests that start and stop the server. This
makes automated testing a bit easier. Additionally, request-sync
was dropped for just request, because it turns out asynchronous
methods are needed so that the calls don't block the event loop
for the server to respond.

The calls for rpnow.start() and rpnow.stop() are a bit different
now, mainly to include the "quiet" option which makes it so it
doesn't log anything to the console.