Added export page to frontend; removed stats page
Removed the stats page placeholder and icon. It will not be used
in this GUI.

The export page has been added. In addition, it provides the user
an option as to whether or not OOC messages should be included in
the exported text. The API has been updated to include this, and
it has been documented on the REST endpoints sheet.