Wrote all of the planned API test suites
Test suite now fully launches the server, tests starting and
stopping and starting again, tests creating an RP, creating
characters, and sending messages, as well as checking for
updates, and expecting some 400 Bad Request messages for bad API
calls.

The tests could definitely be improved in the future. Many of them
only check the status code and don't do any additional checks for
if the correct content was returned or recorded to the server.
There are also some edge cases not yet tested for. However, this
should be a good start in making sure the API doesn't break
between changes.