Rate limiting, JSON options, API test organization
A lot of loosely related changes in this one.

First, we're using npm module express-rate-limit to enforce the
refresh-milliseconds specification within express. If too many
requests are made within a minute, it will gradually increase the
delay of these request by one-tenth of a second per request. This
way if the rate limit is exceeded only slightly, there won't be
much of a difference, but it will cause noticeable delay in the
event of abuse.

Next, when creating an rpnow server, options can be passed in
by using a JSON object. One of these is the rateLimit option.
This is needed because the Jasmine tests need to turn off rate
limits most of the time so they can run reasonably quickly.

api.js was slightly modularized so that options could be passed
to it from the server.

Finally, I've broken up the main api v1 test suite into two
suites - one which just does basic code coverage, and one which
tests the boundary and error conditions of character and message
posting. The latter will look forward to some expansion soon.