Refactoring in server.js
Console logging has been moved out of server.js and put instead in
app.js. This has removed a lot of `if(!quiet) console.log("...")`
statements which were kind of noisy and confusing.

Additionally, the "this" context in module.exports.start/stop was
confusing, so instead the server and its options are stored in a
simple var statement.

Creating the options object is also much cleaner-looking now.