Technically uses MongoDB now (But it's pretty bad)
Eventually this project will use MongoDB in a way that is actually
good. For now, just to get started, it will write out all of the
RPs to a single BSON document in the database every 5 seconds. I
warned you that it would be bad! But still, it does have some
level of persistence now.

It uses the mongojs package to communicate to Mongo in Node. I
also added the data/ folder to the .gitignore file, since that's
where data will be stored.