Beginning to use Socket.IO instead of REST calls
What if we use Socket.IO for all the server stuff? This, the first
in an experimental series of developments, will explore that idea.

Restructured some stuff in server.js and api.js to work in the
Socket.IO stuff. Socket.IO now delivers the initial load of RP
information upon connecting. Later it will deliver everything!