Keeping consistent data format from db to client
From the way it's stored in MongoDB to the way it's presented to
the client, an RP is very nearly the same JSON throughout. This
broke some of the things on the client end, which is fine because
the frontend is all going to be redone anyway. For isntance,
saving which character you've been using between browser refreshes
no longer works. Additionally, you can't see the messages and
characters that you yourself added until after you refresh the
page. This will be fixed soon. A lot of refactoring came with this
and I don't know if I can adequately explain it all.