Moving from NodeJS 4 to 6
Beginning to migrate from NodeJS 4 to 6, to start making use of
ES6 features (such as the `let` keyword, and others). Replaced
all `var` with `let` in server/app.js and server/server.js, with
no apparent ill effects. Changed the Dockerfile to build off the
Node 6 image instead of Node 4.