Adjusting Dockerfile to streamline development
First, using the default node user for the node container, and
also reorganized the statements so that npm isn't run until the
end. Also clean the npm cache due to the next thing.

Node_modules in the rpnow container is a docker-created volume, so
that it can persist between builds. Docker can decide where to put
it, so I don't have to worry about it.

I'm also specifying to put the server and static folders as
volumes in the container, so that any changes I make to those
files will be updated in the container. This if for development
purposes only, though!