Load RP messages from Angular controller
The new RP UI now uses AngularJS to display messages on the page.
For now, all data is still placeholder data and not actually
fetched from the server.

rp.js and rp.css will be redone to be used in the new UI, but the
old files will be kept for now.