Alignment and responsiveness nitpicks
In a medium sized screen, the RP title, the messages, and the
message box will align in a single 600px column. The other UI
components are also lined up properly. The send button is also
much larger and brighter.

In a small screen, the icons in the top right collapse into a
single down-arrow menu, and the elements squish together as much
as they can.