Paper-like aesthetic overhaul
Although we're still using the ngMaterial components, we're now
using our own styles to make it a bit more flat-looking. We use
solid 1px outlines rather than shadows and bookish fonts to evoke
ink on paper. Also improved alignment and stuff.

A lot of these styles are currently inline. It definitely needs
refactoring but for now I just want it to look good.