Implemented RP download button
Added the button for downloading the RP. Since the entire RP
exists on the client, the browser itself generates the text file
and offers it as a "download" to the user.

I implemented my own line wrap function because I felt embarrassed
about wanting to copy somebody else's. Turns out it's a bit of a
challenge...