Basic .docx generation and downloading
Users can now download a .docx file containing the contents of
their RP. It uses DocxTemplater (which in turn uses JSZip). It
works by taking in a template file and the RP data, and outputting
a filled-in RP document. This is all done within the browser, just
like the .txt file.

The current .docx template is not fully fleshed out; it fails to
incorporate things like character names or colors and could stand
for improvement. A new, better template will probably be created
later.