Added rudimentary character selector
There's now a dropdown menu above the character box that lets you
choose whether you want to post as narrator, OOC, or some
character. It also no longer sends a message as a random non-
existant character; it actually sends with the one you want it to.

The character list is currently populated with dummy characters
that don't exist. A character creator will be added soon.