More advanced logic for how many messages to show
When the user is glued to the bottom of the scrolling story panel,
the screen will only show the last 100 messages. However, if the
user has scrolled up, it will tolerate a total of 200 messages.

This is so that when a user posts, another user viewing the top of
the RP won't see the messages start to disappear immediately; only
when they've scrolled back down to the bottom.