Prettier, more capable character creator
Aside from aesthetic improvements to the character creator, there's
two new important functionalities. First, we're using the
angular-color-picker library to let users choose colors. This seems
like a good, simple color picker. I've also made it so the OK
button changes to the color selected, so the user knows what color
they're giving their character for sure.

Note that we're using version 1.0.10-rpn1 of the color picker. This
is my own fork of the angular-color-picker library which is touch-
enabled and a more appropriate size for the dialog it fits in.