Persist some settings inside of localStorage
We use this localStorageService to bind some parts of our model to
localStorage. That way, even after closing the browser and opening
it again, we still keep our saved settings.

Some settings are scoped to the current RP (ie the state of the
message input box) while other things are scoped globally to RPNow
(sidenav options, the "keep OOC messages" checkbox in the download
menu, and the last color used to create a character.)

If localStorage is not supported, then the settings will go back
to their defaults each time the page reloads. All browsers that
RPNow supports do have localStorage anyway, I believe.

Because we don't want to persist the character dropdown in an
"Add New Character" state, we immediately set a timeout of 0 for
it to change back to its original value. This might be better
behavior anyway, and removes addCharaBox.lastVoice from our model.

On a related note, msgBox.selected wasn't doing anything, and has
been removed.