Enhanced character list styles
We break the "Add new character" button out to the side of the
dropdown when the screen is larger than xs. Also, we have colored
icons next to each character, including different icons for
narrator and ooc.

The menu also looked kind of weird in dark mode so I fixed it with
my own custom styles.