Dropdown voice menu now shows most recent charas
Rather than listing all the characters like the "all characters"
menu does (thereby making it cumbersome and redundant) we now only
show the top 5 most recently used characters there, mixing in the
narrator and OOC too. On small screens, we still have the "all
characters" button there (although if there's no characters yet,
it will just show the "add new" button instead.)

I'm actually storing this not as an array but as a string, which
can be accessed as an array through a function call. This was only
done so that it would be possible for the localStorageService to
remember it.