Add rp page to ngRouter
We've moved the rp page into template files and linked to the
router. A few changes were also needed to make this possible:
the ng-view is now on the body instead of an inner div, and the
various darkmode-related directives were also attached to that
body. Also, the dependencies for the rpnow module are now in
index.html.