Restructuring project
All code is now in /src, except for tests which remain in /spec.

Changing /src/static to /src/www.

app.js was chanaged to the more conventional entrypoint name,
index.js. All other server-side js was moved into /src/server.

index.html is in the /src root like index.js but I'm not sure if
I'll keep it this way. Maybe putting it in /www/app would be
better? I'm leaving it this way for now though.