Serve error JSON to client for bad requests
If the client makes a bad request, the server will serve a JSON
object with an "error" field, rather than sending no response at
all.

If the client requests a non-existent RP, it will now log the
error in the console. TODO: let the user know that the URL they
requested was not valid.