Refactored sockets/API & fixed SPA bug
RP creation is now done through a socket.io call instead of a REST
call. This means that the extra express router for the API can
completely go away.

Introduced the "exit rp" socket.io message (and changed "join rp"
to "enter rp" to mirror it.) The exit message is sent when
navigating away from the RP page, but remaining in the SPA. Doing
this, rather than closing the socket, fixes the bug where
backing out of/going back to the RP page within the app would break
the page.