Move rp downloading into service, fix docx bug
Created the saveRpService to separate out all the RP downloading
code, which was a large portion of the rpController.

Also fixed an issue where the docx download would always include
OOC messages even if the user unchecks the "include OOC" checkbox.

One more thing: the .docx template is cached in the browser once
fetched so the browser won't fetch it twice.