Begin reintroducing spec for rpnow server
We're using mostly the same logic as before to test starting and
stopping the server, and then checking to make sure that the
single-page application's index.html is loading properly. Also
that static resources are still available.

Since the API has changed completely from REST calls to socket.io
messages, that part is disabled for now, to be redone.