Add basic tests for socket.io comms
Similar to the old spec, we just use socket.io to send some basic
messages to create an RP, add messages and a chara, then leave.
More thorough testing is needed for things like finding out if
other clients receive the messages sent, and if the messages
persist after refreshing.