Store charaId instead of chara obj in msg
Rather than storing an entire chara with a name and color
in each message, we just use charaId to reference its index in the
charas array. This is to ensure forwards compatibility with chara
edits, which should affect previous messages made by that chara.
It also happens to take up less space which is a nice bonus.