Pretty title page + css refactoring
The title page now looks great with some css and a couple angular
material components.

While making this, I refactored a lot of CSS. now, index.css
contains common css for the site, while rp.template.css has
css specific to the rp page. message.css has all the css for the
rp messages; it's huge so it's separate, plus it could be used on
the reader page too later. Finally, I got the canonical
material-icons.css and put it in /lib, so it wasn't mixed up in
my own css.

Also fixed a bug where rolling a title didn't immediately apply
the new title text to the page, so the title that displayed wasn't
actually the final one that was rolled. Now it's working as
expected.