Eliminate Dockerfile, just use node image+volumes
I read something online (I wish I remember where!) about the idea
of Docker containers not being applications, but pieces of
infrastructure. I really liked that idea! So now we're not building
a separate rpnow image anymore; we're just using the default node
image which uses some volumes for the source files, and running an
npm script to install the node_modules and launch the server.