Dockerized build process with Gulp.js
We're now using Gulp to automate building the frontend of the site. The
build process doesn't do much yet; it just does babel for app.js and
minifies the css.

The Gulp build process is also dockerized. A new fourth container builds
the frontend, stores it in a shared volume, and exits. Because of this,
we no longer need to store the build in the repo.

Side note: for the two node-based images, I had to settle for using the
root user to run the node scripts, because unlike I didn't realize that
docker-compose volumes are automatically set up with root permissions.
I'm not sure how this wasn't an issue before but it has been one now. I
spent a lot of time trying to automatically change those permissions but
it's really not worth the benefits right now, I don't think.