Dockerize unit tests
Added docker-compose.tests.yml which does NOT extend the default
dockerfile. This opens an ephemeral database image and runs the
jasmine tests on the API server.

Moved docker-compose.production.yml to docker-compose.production.override.yml,
to add the distinction that the production mode does indeed override
the base docker-compose file, while docker-compose.tests.yml does not.

Because every part of development is now dockerized, node_modules
should no longer appear in the repo, and so the .gitignore is no
longer needed.