Trust NGINX proxy to provide the client's IP
Because Node is now behind NGINX as a proxy, the normal IP reported by
socet.io will be incorrect. Use the x-forwarded-for header instead, but
fall back to the old IP if it's not available for some reason.

socket.io packet logging has been improved to make it easier to see this
IP (and other information) in the server's logs.