Better handling of message dropping; refactoring
Refactored all of the RP socket.io client logic into its own
angular factory, providing better separation of concerns.

Also, added a "message send queue" to that factory, which maintains
a list of messages that haven't been sent yet. Upon a reconnect
event, the factory realizes that the messages probably need to be
sent again due to connection issues, and so it sends them again.

If a chara fails to send... it just stays on the spinning wheel
forever. This could definitely be improved later.