Add image posts
You can now post externally hosted images to an RP. Later I may
implement an upload feature or integrate with imgur, but right now
the user can just paste the image URL into a dialog.

The API does a quick check to make sure the URL exists and that
the content-type is some kind of image. Perhaps more checking
would be appropriate but for now I'm not really sure what that
would entail.

Added styling for the image post, as well as updating the RP
download features to work with the images. (Right now, they both
just show the URL of the image. In the future, the .docx might
display the image inline, but that's for later.)

TODO: add unit tests for image posts.