Aesthetic improvements to msg box, and tooltips
Using md-tooltips for some icon buttons and things where function
might not be immediately obvious. They are ng-if'd with a hasMouse
variable which is set to true if the browser detects a mousemove
event.

Removed the md-select from the character box title and just made
it into a button which immediately opens up the character drawer
instead, since that has all the recent characters on it anyway.

Because the "recent characters" menu is now only in the drawer,
I made it so it only shows user-made characters and not the
narrator or OOC voices, since those are both always at the very
top anyway.

Other misc aesthetic changes for a cleaner look overall.