Message content editing (needs refining & testing)
Once you've successfully sent a message, that message will have an
icon button which switches that message to an editing mode. The
user can either save the edit or cancel it.

Some crypto is used to secure this. A random password is given to
the message sender in the "add message" callback, and its hash is
stored in the database. Upon attempting to edit the image, the API
verifies that the password does indeed resolve to the expected
password hash. This is all done transparently to the user.

Because these random passwords are only stored in the client-side
javascript, the client can no longer edit their sent messages
after reloading the page later. This is by design.

Although existing tests are updated to accomodate "add message"
packets including secrets and hashes, there are not any tests to
actually verify that editing works as it should. This should be
fixed soon.