Increased resilience in production
The live server was running into an issue where node and/or the
database would run out of memory, and those containers would crash,
then would automatically restart. This is all well and fine, except
the web server would fail to update the IP address to which it
needed to forward API requests. Setting the resolver should fix
that issue.

Also, if the build container fails to complete, it will try again
up to 3 times before giving up.