Let users edit their own messages forever
The RP client fetches a secret/hash pair from the server via a REST
call, which it then uses for all future messages by storing it in
localStorage. The hash is sent as part of the message. Instead of
one secret per message, it's now one secret per client per RP.