Migrate some of the trickier libraries into bower
The socket.io-client package on bower doesn't have a bower.json so
we have to manually specify which files to include.

Because changes have been made to angular-color-picker, I had to
fork the repository on GitHub and reference that branch in our
bower.json.

angular-scroll-glue just had to be tested because I thought I had
made changes to it, but ultimately I'm pretty sure I didn't.