Switch from angular-route to UI-router
The main motivation for doing this was because I suspected it would
help fix the issue with the $templateCache not working in the prod
build. This turned out to be correct; I got it to work now.

A secondary motivation is that I suspect that implementing the
archive reader would benefit from UI-router, because it has nested
views. Time will tell if this is true, but it probably won't hurt.

Using a fork of ngMeta because the main branch does not support
UI-router 1.0 and above.