WIP: moving event broadcasting into the API's model layer
Rather than the socket.io layer keeping track of when messages are received and then re-broadcasting them, we have the model layer do this. This means we could probably support multiple ways for messages/charas to be sent to the server, IE by rest calls or some other streaming system.

Right now there's a problem - the model layer is not aware of the identity of senders, and socket.io doesn't remember either, so it broadcasts it to everyone, even the sender. Clients will therefore see their own messages twice until they refresh the page, which may cause further problems. This needs to be fixed soon.